Hisilicon/Sm750Dxe: Unify binary and fix GOP frame buffer base
Sm750Dxe is a generic PCIe device driver for SM750 VGA device, so it
is not necessary to maintain two different binary images for D03 and
D05 respectively.

The main difference between D03 and D05 is the implementation of
SerialPortLib, which causes BaseDebugLibSerialPort to be different. So
we switch to null DebugLib for release build and get a unified
Sm750Dxe binary.

The code in SM750 driver treated the address returned from
PciIo->GetBarAttributes() as device address; this should be fixed
after edk2 commit dc080d3 since GetBarAttributes() returns host
address from then on.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Heyi Guo <heyi.guo@linaro.org>
Signed-off-by: Yi Li <phoenix.liyi@huawei.com>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Leif Lindholm <leif.lindholm@linaro.org>
Cc: Michael D Kinney <michael.d.kinney@intel.com>